+++ spec/spec_helper.rb
@@ -9,6 +9,7 @@ end
 
 require 'faker'
 require 'rspec'
+require 'rspec/retry'
 require 'onlyoffice_api'
 require_relative 'spec_helper/portal_cleanup'
 require_relative 'spec_helper/request_helper'
@@ -28,10 +29,13 @@ end
 
 configure_test_portal
 
-RSpec.configure do |c|
-  c.before(:all) do
+RSpec.configure do |config|
+  config.before(:all) do
     @data_collector = {}
   end
+
+  config.default_retry_count = 3
+  config.verbose_retry = true
 end
 
 shared_examples_for 'an api request' do |*flags|